code style fix & memory leak fix

+++ src/CSV/parcer.h
@@ -2,4 +2,4 @@
 
 #include <stdio.h>
 
-int columnWidth(int **maxColumns, int *currentCapacity, FILE *file);
+int columnWidth(int** maxColumns, size_t* currentCapacity, FILE* file);
